+++ pkg/eval_query_test.go
@@ -101,8 +101,8 @@ func TestMacrosBuilder(t *testing.T) {
 			"$perSecond",
 			"/* comment */\n$perSecond(from_total, from_amount) FROM requests",
 			"/* comment */\nSELECT t,"+
-				" if(runningDifference(max_0) < 0, nan, runningDifference(max_0) / runningDifference(t/1000)) AS max_0_Rate,"+
-				" if(runningDifference(max_1) < 0, nan, runningDifference(max_1) / runningDifference(t/1000)) AS max_1_Rate"+
+				" if(runningDifference(max_0) < 0, nan, runningDifference(max_0) / runningDifference(t/1000)) AS max_0_PerSecond,"+
+				" if(runningDifference(max_1) < 0, nan, runningDifference(max_1) / runningDifference(t/1000)) AS max_1_PerSecond"+
 				" FROM ("+
 				" SELECT $timeSeries AS t,"+
 				" max(from_total) AS max_0,"+
@@ -114,14 +114,46 @@ func TestMacrosBuilder(t *testing.T) {
 			q.perSecond,
 		),
 		newMacrosTestCase(
+			"$delta",
+			"/* comment */\n$delta(from_total, from_amount) FROM requests",
+			"/* comment */\nSELECT t,"+
+				" runningDifference(max_0) AS max_0_Delta,"+
+				" runningDifference(max_1) AS max_1_Delta"+
+				" FROM ("+
+				" SELECT $timeSeries AS t,"+
+				" max(from_total) AS max_0,"+
+				" max(from_amount) AS max_1"+
+				" FROM requests"+
+				" WHERE $timeFilter"+
+				" GROUP BY t"+
+				" ORDER BY t)",
+			q.delta,
+		),
+		newMacrosTestCase(
+			"$increase",
+			"/* comment */\n$increase(from_total, from_amount) FROM requests",
+			"/* comment */\nSELECT t,"+
+				" if(runningDifference(max_0) < 0, 0, runningDifference(max_0)) AS max_0_Increase,"+
+				" if(runningDifference(max_1) < 0, 0, runningDifference(max_1)) AS max_1_Increase"+
+				" FROM ("+
+				" SELECT $timeSeries AS t,"+
+				" max(from_total) AS max_0,"+
+				" max(from_amount) AS max_1"+
+				" FROM requests"+
+				" WHERE $timeFilter"+
+				" GROUP BY t"+
+				" ORDER BY t)",
+			q.increase,
+		),
+		newMacrosTestCase(
 			"$perSecondColumns",
 			"/* comment */\n$perSecondColumns(concat('test',type) AS from_alias, from_total) FROM requests WHERE type IN ('udp', 'tcp')",
 			"/* comment */\nSELECT t,"+
-				" groupArray((from_alias, max_0_Rate)) AS groupArr"+
+				" groupArray((from_alias, max_0_PerSecond)) AS groupArr"+
 				" FROM ("+
 				" SELECT t,"+
 				" from_alias,"+
-				" if(runningDifference(max_0) < 0 OR neighbor(from_alias,-1,from_alias) != from_alias, nan, runningDifference(max_0) / runningDifference(t/1000)) AS max_0_Rate"+
+				" if(runningDifference(max_0) < 0 OR neighbor(from_alias,-1,from_alias) != from_alias, nan, runningDifference(max_0) / runningDifference(t/1000)) AS max_0_PerSecond"+
 				" FROM ("+
 				" SELECT $timeSeries AS t,"+
 				" concat('test', type) AS from_alias,"+
@@ -136,6 +168,54 @@ func TestMacrosBuilder(t *testing.T) {
 				" GROUP BY t"+
 				" ORDER BY t",
 			q.perSecondColumns,
+		),
+		newMacrosTestCase(
+			"$deltaColumns",
+			"/* comment */\n$deltaColumns(concat('test',type) AS from_alias, from_total) FROM requests WHERE type IN ('udp', 'tcp')",
+			"/* comment */\nSELECT t,"+
+				" groupArray((from_alias, max_0_Delta)) AS groupArr"+
+				" FROM ("+
+				" SELECT t,"+
+				" from_alias,"+
+				" if(neighbor(from_alias,-1,from_alias) != from_alias, 0, runningDifference(max_0)) AS max_0_Delta"+
+				" FROM ("+
+				" SELECT $timeSeries AS t,"+
+				" concat('test', type) AS from_alias,"+
+				" max(from_total) AS max_0"+
+				" FROM requests"+
+				" WHERE $timeFilter"+
+				" AND type IN ('udp', 'tcp')"+
+				" GROUP BY t, from_alias"+
+				" ORDER BY from_alias, t"+
+				")"+
+				")"+
+				" GROUP BY t"+
+				" ORDER BY t",
+			q.deltaColumns,
+		),
+		newMacrosTestCase(
+			"$increaseColumns",
+			"/* comment */\n$increaseColumns(concat('test',type) AS from_alias, from_total) FROM requests WHERE type IN ('udp', 'tcp')",
+			"/* comment */\nSELECT t,"+
+				" groupArray((from_alias, max_0_Increase)) AS groupArr"+
+				" FROM ("+
+				" SELECT t,"+
+				" from_alias,"+
+				" if(runningDifference(max_0) < 0 OR neighbor(from_alias,-1,from_alias) != from_alias, 0, runningDifference(max_0)) AS max_0_Increase"+
+				" FROM ("+
+				" SELECT $timeSeries AS t,"+
+				" concat('test', type) AS from_alias,"+
+				" max(from_total) AS max_0"+
+				" FROM requests"+
+				" WHERE $timeFilter"+
+				" AND type IN ('udp', 'tcp')"+
+				" GROUP BY t, from_alias"+
+				" ORDER BY from_alias, t"+
+				")"+
+				")"+
+				" GROUP BY t"+
+				" ORDER BY t",
+			q.increaseColumns,
 		),
 	}
 	r := require.New(t)